modify default datatype as int16 for new empty image

+++ froi/gui/component/datamodel.py
@@ -309,7 +309,7 @@ class VolumeListModel(QAbstractListModel):
         """
         if data is None:
             new_data = np.zeros(self._data[0].get_data_shape()[0:3],
-                                dtype=np.uint8)
+                                dtype=np.int16)
         else:
             new_data = data
         new_header = self._data[0].get_header().copy()
